ditch readline for now

+++ src/index.ts
@@ -1,34 +1,29 @@
 import * as fs from 'fs';
-import * as readline from 'readline';
 import * as commandLineArgs from 'command-line-args';
 
 const options = commandLineArgs([
-    { name: 'out', alias: 'o', type: directory => ({ directory, exists: fs.existsSync(directory) }), defaultOption: true },
+    { name: 'file', alias: 'f', type: file => ({ name: file, exists: fs.existsSync(file)}), defaultOption: true},
+    { name: 'out', alias: 'o', type: directory => ({ name: directory, exists: fs.existsSync(directory) }) },
 ]);
 
 const SUPPORTED_LANGUAGES = 'ts|py';
 
-let input = '';
+const { out, file } = options;
 
-const rl = readline.createInterface({
-    input: process.stdin,
-    output: process.stdout,
-    terminal: false,
-});
+if (!file.exists) {
+    console.log(`Couldn't open file ${file.name}`)
+    process.exit(1);
+}
 
-rl.on("line", (line: string) => {
-    input += `${line}\n`;
-})
+const input = fs.readFileSync(file.name).toString();
 
-rl.on('close', () => {
-    const re = new RegExp(`^\`\`\`(${SUPPORTED_LANGUAGES})\n(.*?)\n\`\`\`\n$`, 'gims');
-    let counter = 0;
-    for (let match = re.exec(input); match; match = re.exec(input)) {
-        fs.writeFileSync(
-            `${options.out.directory}/${counter}.${match[1]}`,
-            match[2]
-        );
-        console.log(`created ${counter}.${match[1]}`);
-        counter++;
-    }
-});
+const re = new RegExp(`^\`\`\`(${SUPPORTED_LANGUAGES})\n(.*?)\n\`\`\`\n$`, 'gims');
+let counter = 0;
+for (let match = re.exec(input); match; match = re.exec(input)) {
+    fs.writeFileSync(
+        `${out.name}/${counter}.${match[1]}`,
+        match[2]
+    );
+    console.log(`created ${counter}.${match[1]}`);
+    counter++;
+}
